Add /nav/:editionId.json endpoint to allow lazy-nav loading in DCR
Motivated by sticky nav tests in DCR. So important to revert
changes if tests end up unsuccessful.

Note, as part of this, DCR nav data has been moved into the
navigation package of common.
