+++ onward/conf/routes
@@ -78,6 +78,7 @@ GET        /business-data/stocks.json                 controllers.StocksControll
 
 # DCR
 GET        /story-package/*path.json                  controllers.StoryPackageController.render(path)
+GET        /nav/:editionId.json                     controllers.NavigationController.renderDCRNav(editionId: String)
 
 # AMP
 GET        /most-read-mf2.json                        controllers.MostPopularController.renderPopularMicroformat2()
